+++ api/src/main/java/uob_todo/api/exceptions/BadRequestException.java
@@ -1,4 +1,4 @@
-package uob_todo.exceptions;
+package uob_todo.api.exceptions;
 
 public class BadRequestException extends Exception {
     public BadRequestException(String message) {
