+++ api/src/test/java/uob_todo/TodoControllerTest.java
@@ -8,6 +8,8 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
+import uob_todo.api.TodoItem;
+import uob_todo.api.TodoRepository;
 
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
@@ -26,7 +28,7 @@ public class TodoControllerTest {
 
     @Test
     public void getAllTodoItems__empty() throws Exception {
-        this.mvc.perform(get("/todos").accept(MediaType.APPLICATION_JSON))
+        this.mvc.perform(get("/api/todos").accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andExpect(content().json("[]"));
     }
@@ -38,14 +40,14 @@ public class TodoControllerTest {
         todoSource.save(new TodoItem("Bop it"));
         todoSource.save(new TodoItem("Twist it"));
 
-        this.mvc.perform(get("/todos").accept(MediaType.APPLICATION_JSON))
+        this.mvc.perform(get("/api/todos").accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andExpect(content().json("[{\"title\":\"Bop it\",\"completed\":false},{\"title\":\"Twist it\",\"completed\":false}]"));
     }
 
     @Test
     public void getTodo__notfound() throws Exception {
-        this.mvc.perform(get("/todos/99999").accept(MediaType.APPLICATION_JSON))
+        this.mvc.perform(get("/api/todos/99999").accept(MediaType.APPLICATION_JSON))
                 .andExpect(status().is(404))
                 .andExpect(content().json("{\"error\":\"item not found\"}"));
     }
